+++ README.md
@@ -3,11 +3,5 @@ init-repository
 ====
 [![Build Status](https://travis-ci.org/hmrc/init-repository.svg?branch=master)](https://travis-ci.org/hmrc/init-repository) [ ![Download](https://api.bintray.com/packages/hmrc/releases/init-repository/images/download.svg) ](https://bintray.com/hmrc/releases/init-repository/_latestVersion)
 
-This tool automatically creates a repository for the new Tax Platform. Specifically it will:
-
-* Create a new GitHub repository in [HMRC on GitHub](https://www.github.com/hmrc) assigned to a given team
-* Create a new Bintray releases package in [HMRC on Bintray](https://www.bintray.com/hmrc)
-* Create a new Bintray release-candidates package [HMRC on Bintray](https://www.bintray.com/hmrc)
-
-The HMRC and Bintray credentials are intentionally omitted from this repository for security reasons. This repository is useless without them.
+This tool automatically creates a repository in [HMRC on GitHub](https://www.github.com/hmrc) and assigns it to a given team
