+++ build.sbt
@@ -1,0 +1,15 @@
+import sbtassembly.AssemblyKeys.assembly
+import sbt._
+
+val appName = "init-repository"
+
+lazy val library = Project(appName, file("."))
+  .enablePlugins(SbtAutoBuildPlugin, SbtGitVersioning, SbtArtifactory)
+  .settings(
+    majorVersion                     := 1,
+    makePublicallyAvailableOnBintray := true,
+    libraryDependencies ++= AppDependencies.compile ++ AppDependencies.test,
+    AssemblySettings(),
+    addArtifact(artifact in (Compile, assembly), assembly)
+  )
+
